Do not display request if there's a convo for it
Requests get deleted after accepting it, but in cases where the request
might be present (bad propagation or accepting a request failed
(it is not atomic), both the request and conversation will show up, so
now the request is hidden in UI if there's a convo for it. This does
NOT hide requests by the same user, just the request that corresponds
to an existing convo.